# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Core server.",
        "title": "GoCore Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DuongVu",
            "url": "http://www.swagger.io/support",
            "email": "vuduongcalvin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/countries": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get list country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Get list country",
                "operationId": "countriesList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JSON string of filter. E.g: {\"field_name\":\"value\"}",
                        "name": "f",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Number of records per page",
                        "name": "l",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "default": "ASC",
                        "description": "Sort direction order must be one of ASC, DESC",
                        "name": "o",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current page number",
                        "name": "p",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field name for sorting",
                        "name": "s",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/country.ListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creates new country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Creates new country",
                "operationId": "countriesCreate",
                "parameters": [
                    {
                        "description": "CreationData",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/country.CreationData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AuthToken"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    }
                }
            }
        },
        "/v1/countries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns a single country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Returns a single country",
                "operationId": "countriesView",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Country"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Deletes an country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Deletes an country",
                "operationId": "countriesDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggOKResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Updates country information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Updates country information",
                "operationId": "countriesUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateData",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/country.UpdateData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swaggerutil.SwaggErrDetailsResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "country.CreationData": {
            "type": "object",
            "required": [
                "code",
                "name",
                "phone_code"
            ],
            "properties": {
                "code": {
                    "description": "example: vn",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                },
                "name": {
                    "description": "example: Vietnam",
                    "type": "string",
                    "minLength": 3
                },
                "phone_code": {
                    "description": "example: +84",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                }
            }
        },
        "country.ListResp": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "example: [{\"id\": 1, \"created_at\": \"2020-01-14T10:03:41Z\", \"updated_at\": \"2020-01-14T10:03:41Z\", \"name\": \"Singapore\", \"code\": \"SG\", \"phone_code\": \"+65\"}]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Country"
                    }
                },
                "total_count": {
                    "description": "example: 1",
                    "type": "integer"
                }
            }
        },
        "country.UpdateData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "example: vn",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                },
                "name": {
                    "description": "example: Vietnam",
                    "type": "string",
                    "minLength": 3
                },
                "phone_code": {
                    "description": "example: +84",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                }
            }
        },
        "model.AuthToken": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "model.Country": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "description": "The time that record is created",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the record",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone_code": {
                    "type": "string"
                },
                "updated_at": {
                    "description": "The latest time that record is updated",
                    "type": "string"
                }
            }
        },
...
        "server.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "swaggerutil.SwaggErrDetailsResp": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/server.HTTPError"
                }
            }
        },
        "swaggerutil.SwaggOKResp": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}